+++ library/src/main/java/company/tap/gosellapi/internal/api/responses/BINLookupResponse.java
@@ -41,6 +41,10 @@ public final class BINLookupResponse implements BaseResponse {
     @Expose
     private String country;
 
+    @SerializedName("card_type")
+    @Expose
+    private String cardType;
+
     /**
      * Defines if a card with the given bin_number requires address fields to be passed to
      * Charge API.
@@ -102,4 +106,14 @@ public final class BINLookupResponse implements BaseResponse {
     public String getCountry() {
         return country;
     }
+
+    /**
+     * Gets cardType.
+     *
+     * @return cardType.
+     */
+    public String getCardType() {
+        return cardType;
+    }
+
 }