+++ library/src/main/java/company/tap/gosellapi/internal/api/responses/PaymentOptionsResponse.java
@@ -49,7 +49,9 @@ public final class PaymentOptionsResponse implements BaseResponse {
     @Expose
     @Nullable private ArrayList<SavedCard> cards;
 
-
+    @SerializedName("settlement_currency")
+    @Expose
+    @Nullable private String settlement_currency;
     /**
      * Gets id.
      *
@@ -117,5 +119,16 @@ public final class PaymentOptionsResponse implements BaseResponse {
 
         return cards;
     }
+
+    /**
+     * Gets settlement_currency.
+     *
+     * @return the settlement_currency
+     */
+    @Nullable
+    public String getSettlement_currency() {
+        return settlement_currency;
+    }
+
     ////////////////////////////////////////////////////////////////////////////////////////////
 }