+++ library/src/main/java/company/tap/gosellapi/open/controllers/ThemeObject.java
@@ -68,6 +68,16 @@ public class ThemeObject {
         private int                     payButtonTextSize;
 
 
+        /**
+         * Dialog Appearance
+         */
+
+        private int                     dialogTextColor;
+        private int                     dialogTextSize;
+        private int                     dialogBackgroundColor;
+        private int                     dialogbuttonColor;
+
+
         ////////////////////////////////////////////////////////////////////////////////////////////
 
         /**
@@ -88,7 +98,40 @@ public class ThemeObject {
                 return this;
         }
 
+        ////////////////////////////////////////////////////////////////////////////////////////////
+        /**
+         * dialogTextColor
+         * @param dialogTextColor
+         */
+        public ThemeObject setDialogTextColor(int dialogTextColor) {
+                this.dialogTextColor = dialogTextColor;
+                return this;
+        }
+        /**
+         * dialog  TextSize
+         * @param dialogTextSize
+         */
+        public ThemeObject setDialogTextSize(int dialogTextSize) {
+                this.dialogTextSize = dialogTextSize;
+                return this;
+        }
+        /**
+         * dialog backgroundColor
+         * @param dialogBackgroundColor
+         */
+        public ThemeObject setDialogBackgroundColor(int dialogBackgroundColor) {
+                this.dialogBackgroundColor = dialogBackgroundColor;
+                return this;
+        }
+        /**
+         * dialog buttonColor
+         * @param dialogbuttonColor
+         */
+        public void setDialogbuttonColor(int dialogbuttonColor) {
+                this.dialogbuttonColor = dialogbuttonColor;
+        }
 
+        ////////////////////////////////////////////////////////////////////////////////////////////
 
         /**
          * Translucent color
@@ -527,6 +570,36 @@ public class ThemeObject {
                 return this.payButtonTextSize;
 
                 return 14;
+        }
+        /**
+         *
+         * @return Dialog  TextColor
+         */
+        public int getDialogTextColor() {
+                return dialogTextColor;
+        }
+        /**
+         *
+         * @return Dialog TextSize
+         */
+
+        public int getDialogTextSize() {
+                return dialogTextSize;
+        }
+        /**
+         *
+         * @return Dialog BackgroundColour
+         */
+
+        public int getDialogBackgroundColor() {
+                return dialogBackgroundColor;
+        }
+        /**
+         *
+         * @return DialogbuttonColor
+         */
+        public int getDialogbuttonColor() {
+                return dialogbuttonColor;
         }